Move withdraw functions to their own service
